survival_time/requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
autopep8==2.0.0
blinker==1.4
cachetools==4.2.2
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==22.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
Cython==0.29.14
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
gast==0.3.3
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.38.0
h5py==2.10.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
jedi==0.17.2
Jinja2==2.10.1
joblib==1.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.0
more-itertools==4.2.0
mpi4py==3.0.3
neovim==0.3.1
netifaces==0.10.4
numpy==1.20.2
oauthlib==3.1.0
openslide-python==1.1.2
opt-einsum==3.3.0
pandas==1.2.4
parso==0.7.1
pexpect==4.6.0
Pillow==8.2.0
pluggy==0.13.1
protobuf==3.17.3
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.9.1
pycups==1.9.73
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pypng==0.0.20
PyQRCode==1.2.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-snappy==0.5.3
PyTrie==0.2
pytz==2021.1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scipy==1.4.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.3
ssh-import-id==5.10
systemd-python==234
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
tomli==2.0.1
torch==1.9.0+cu111
torch-tb-profiler==0.3.1
torchaudio==0.9.0
torchvision==0.10.0+cu111
Twisted==18.9.0
txaio==2.10.0
typing-extensions==3.7.4.3
u-msgpack-python==2.1
ubuntu-advantage-tools==27.6
ufw==0.36
ujson==4.0.2
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
Werkzeug==2.0.1
wrapt==1.12.1
wsaccel==0.6.2
xkit==0.0.0
zipp==1.0.0
zope.interface==4.7.1
